project-01-adventure-game-zdupre18/src/adventure_game/Character.java
package adventure_game;
import java.util.ArrayList;
/**class Characte */
import adventure_game.items.Consumable;
/** Character takes each player and makes their description including the max health the character will have to the base damage the character goes into combat with 
*/
abstract public class Character{
    private int maxHealth;
    private int health;

    private int maxMana;
    private int mana;

    private int baseDamage;

    private String name;

    private ArrayList<Consumable> items;

    // Character Conditions:
    private int turnsVulnerable; // number of turns Character is vulnerable
    private int turnsInvincible; // number of turns Character takes no damage
    private int turnsStunned; // number of turns Character gets no actions
    // buffer factor for next attack
    // E.g, if 2.0, the next attack will do double damage
    private double tempDamageBuff;
/** creates character class for each character and creates descrtiption
 * @param name characters name 
 * @param health characters health 
 * @param mana characters mana 
 * @param damage characters damage 
  */
    public Character(String name, int health, int mana, int damage){
        this.name = name;
        this.maxHealth = health;    
        this.health = health;
        this.maxMana = mana;
        this.mana = mana;
        this.baseDamage = damage;
        this.tempDamageBuff = 1.0;
        items = new ArrayList<Consumable>();
    }

    /** override StringtoString where and prints out each character description  */
    @Override
    public String toString(){
        String output;
        output = "";
        output += "Name " + getName() + "\n";
        output += "hp " + getHealth() + "\n";
        output += "mana " + getMana() + "\n";
        output += "damage " + getBaseDamage() + "\n";
        return output;
    }

    /**
     * Get the name of this Character
     * @return the name of this Character
     */
    public String getName(){
        return this.name;
    }

    public int getHealth(){
        return this.health;
    }

    public int getMaxHealth(){
        return this.maxHealth;
    }

    public int getMaxMana(){
        return this.maxMana;
    }

    public int getMana(){
        return this.mana;
    }

    public int getBaseDamage(){
        return this.baseDamage;
    }

    public boolean isAlive(){
        return this.health > 0;
    }

    abstract void takeTurn(Character other);
/**  @param other attack set to let character attakc and due damage on opponent unless they defend 
 */
    public void attack(Character other){
        if(other.isInvincible()){
            System.out.printf("%S is unable to attack %S!\n", 
                                this.getName(), 
                                other.getName());

            other.decreaseTurnsInvincible();
            return;
        }
        
        double modifier = Game.rand.nextDouble();
        modifier = (modifier*0.4) + 0.8;
        int damage = (int)(this.baseDamage * modifier);
        // apply temporary damage buff, then reset it back to 1.0
        damage *= this.tempDamageBuff;
        this.tempDamageBuff = 1.0;

        if(other.isVulnerable()){
            damage *= 1.5;
            other.decreaseTurnsVulnerable();
        }

        System.out.printf("%s dealt %d damage to %s\n", 
                            this.getName(), 
                            damage, 
                            other.getName());
        other.modifyHealth(-damage);
    }
/** defends against player by using one turn to defend against opponent attack
 * @param other allows to defend for other character 
*/
    public void defend(Character other){
        double chance = Game.rand.nextDouble();
        if(chance <=0.75){
            System.out.printf("%s enters a defensive posture and charges up their next attack!\n", this.getName());
            this.setAsInvincible(1);
            this.setTempDamageBuff(2.0);
        } else {
            System.out.printf("%s stumbles. They are vulnerable for the next turn!\n", this.getName());
            this.setAsVulnerable(1);
        }
    }
   /**   modifies health if healing potion is taken and also is health is 0 knows to end the game 
   * @param modifier makes sure to modify health   */ 
    public void modifyHealth(int modifier) {
        this.health += modifier;
        if(this.health < 0){
            this.health = 0;
        }
        if(this.health > this.getMaxHealth()){
            this.health = this.getMaxHealth();
        }
    }

    /* CONDITIONS */
    public void setAsVulnerable(int numTurns){
        this.turnsVulnerable = numTurns;
    }

    public boolean isVulnerable(){
        return this.turnsVulnerable > 0;
    }

    public void decreaseTurnsVulnerable(){
        this.turnsVulnerable--;
    }

    public void setAsInvincible(int numTurns){
        this.turnsInvincible = numTurns;
    }

    public boolean isInvincible(){
        return this.turnsInvincible > 0;
    }

    public void decreaseTurnsInvincible(){
        this.turnsInvincible--;
    }

    public void setAsStunned(int numTurns){
        this.turnsStunned = numTurns;
    }

    public boolean isStunned(){
        return this.turnsStunned > 0;
    }

    public void decreaseTurnsStunned(){
        this.turnsStunned--;
    }

    /**
     * Set the temporary damage buffer. 
     * 
     * This is a multiplicative factor which will modify the damage 
     * for the next attack made by this Character. After the next 
     * attack, it will get reset back to 1.0
     * 
     * @param buff the multiplicative factor for the next attack's
     * damage.
     */
    public void setTempDamageBuff(double buff){
        this.tempDamageBuff = buff;
    }
/** @param item estbalishes if consumable item is in the characters grasp  */
    public void obtain(Consumable item){
        items.add(item);
    }

    public void useItem(Character owner, Character other){
        int i = 1;
        System.out.printf("  Do you want to use:\n");
        for(Consumable item : items){
            System.out.printf("    %d: %S\n", i, item.getClass().getName());
            i++;
        }
        System.out.print("  Enter your choice: ");
        int choice = Game.in.nextInt();
        items.get(choice-1).consume(owner);
        items.remove(choice-1);
    }
/** @return items.isEmpty 
 * hasItems to establish use of items
 */

    public boolean hasItems(){
        return !items.isEmpty();
    }
}
